Populate Pokemon Data
use node populate/insert_pokemon in terminal to populate
